add test functions
